LayoutTests/fast/repaint/background-shorthand-with-gradient-and-height-changes-expected.txt

This test verifies that gradient background gets repainted properly after child box height change.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS repaintRects.indexOf('8 84 40 220') is not -1
PASS successfullyParsed is true

TEST COMPLETE
